[Feature] Attribute System
[Details]
Initial push, nothing is needed in the scene.
Use 'PoolMember' to create elements which are affected by attributes.
Use 'Buff' or 'IAttributeSource' to create sources of attributes.
Use 'AttributePool' to create a collection of attribute sources which members can be added to.
Use 'Attribute->Main' in the toolbar to create full sets of attributes; simply input name of the new attribute, and toggle which elements you wish to be created.