[Feature, Bugfix] Objects that are placed on the client now get connected to the clients grid. Fixed various errors when a player disconnected from the lobby
When the server spawns an object on a grid, the object will now also be connected to the clients local version of the grid. This means that the red/green placement indicator should work on the client now
Fixed various problems that happened when a client would disconnect from the lobby, such as the host getting kicked back to the main menu even though they were still in the lobby. This happened whenever someone other than the host would disconnect