[Feature] Added basic phase flow in combat scene
In the combat scene the states now works as follows:
When the scene starts the game is in the initial phase
Initial Phase
- No building allowed
- Players can basicly only move the camera

After 5 seconds the game moves on to the build phase
Build Phase
- Players can now build
- When a player is ready, they can click the "Ready" button to move to the next phase

After either 30 seconds or when all players are ready, the game moves to the Combat phase
Combat Phase
- Players can no longer build
- Players can basicly only move the camera

To move back to the build phase after the combat phase ( or to move to the end phase) you will have to manually click the trigger button on the PhaseController in the scene on the server (host). This will be changed in the future.